massive project restructure
restructure project to be a single page webapp instead of multipage. Make these changes because the project is still in the planning phase so a restructure is viable
