+++ public/scripts/client.js
@@ -1,0 +1,9 @@
+//client interface
+
+// Document ready
+$(() => {
+// events go here
+
+
+});
+// research nodemon docs
