+++ server/index.js
@@ -2,15 +2,13 @@
 require('dotenv').config();
 
 // Web server config
-const sassMiddleware = require('./lib/sass-middleware');
+const sassMiddleware = require('../lib/sass-middleware');
 const express = require('express');
 const morgan = require('morgan');
 const favicon = require('serve-favicon');
-
+const path = require('path');
 const PORT = process.env.PORT || 8080;
 const app = express();
-
-app.set('view engine', 'ejs');
 
 // Load the logger first so all (static) HTTP requests are logged to STDOUT
 // 'dev' = Concise output colored by response status for development use.
@@ -18,7 +16,7 @@ app.set('view engine', 'ejs');
 app.use(morgan('dev'));
 app.use(express.urlencoded({ extended: true }));
 app.use(
-  '/styles',
+  '../styles',
   sassMiddleware({
     source: __dirname + '/styles',
     destination: __dirname + '/public/styles',
@@ -26,7 +24,8 @@ app.use(
   })
 );
 app.use(express.static('public'));
-app.use(favicon(__dirname + '/public/images/favicon.ico'));
+app.use(favicon(path.join(__dirname, '../public', 'images', 'favicon.ico')));
+
 // Separated Routes for each Resource
 // Note: Feel free to replace the example routes below with your own
 const userApiRoutes = require('./routes/users-api');
@@ -44,10 +43,6 @@ app.use('/users', usersRoutes);
 // Home page
 // Warning: avoid creating more routes in this file!
 // Separate them into separate routes files (see above).
-
-app.get('/', (req, res) => {
-  res.render('index');
-});
 
 app.listen(PORT, () => {
   console.log(`Example app listening on port ${PORT}`);
