--- views/partials/header.ejs
@@ -1,5 +1,0 @@
-<header>
-  <nav>
-    <img src="favicon.ico" alt="Logo" width="50" height="50">
-  </nav>
-</header>
